+++ lib/index.js
@@ -1,77 +1,32 @@
-const readline = require("readline");
-const { writeFile } = require("fs/promises");
-const util = require("util");
-const { translateText } = require("./util/translate-text");
+// Mode
+const { interactiveMode } = require("./scripts/interactive-mode");
+const { singleConversionMode } = require("./scripts/single-conversion-mode");
+const { makeSettingFile } = require("./scripts/make-setting-file");
+
+// util
+const { importSetting } = require("./util/import-setting");
 
 module.exports = async () => {
-  const reader = readline.createInterface({
-    input: process.stdin,
-    output: process.stdout,
-  });
-
-  const question = util.promisify(reader.question).bind(reader);
-
-  if (process.argv[2] == "setting" && process.argv[3] == "init") {
-    try {
-      console.log(
-        "Before setting up, you need to issue an API key for Deepl, and apply for API usage on the official Deepl website."
-      );
-      console.log(
-        "If you have not yet issued an API key, you can do so from here. https://www.deepl.com/pro-api"
-      );
-      const apiType = await question(
-        "What is the Type of the API Key?(pro or free): "
-      );
-      const authKey = await question("Please enter your auth key.: ");
-      const targetLang = await question(
-        "Select the language you want to translate into(EN or ZN or JA etc): "
-      );
-
-      let setting = {
-        apiType,
-        authKey,
-        targetLang,
-      };
-
-      const settingJson = JSON.stringify(setting);
-      try {
-        await writeFile(__dirname + "/../setting.json", settingJson);
-      } catch (e) {
-        console.log("writeFile Error");
-        console.log(e);
-      }
-    } catch {
-      console.log("Interrupted. Please try again from the beginning.");
-    } finally {
-      console.log("The setting file has been generated.");
-      process.exit(0);
-    }
-  }
-
   let setting;
 
-  try {
-    setting = require("../setting.json");
-  } catch {
-    console.log("Can't find setting file.");
-    console.log("Please run 'dpl setting init'.");
-    console.log("See Readme for details.");
-    process.exit(0);
-  }
-
-  reader.on("line", async (line) => {
-    const translatedText = await translateText(line, setting);
-
-    console.log(translatedText);
-    reader.prompt();
-  });
-
-  if (process.argv[2]) {
-    const translatedText = await translateText(process.argv[2], setting);
-
-    console.log(translatedText);
-    process.exit(0);
-  } else {
-    reader.prompt();
+  switch (process.argv.length) {
+    case 2:
+      setting = importSetting();
+      interactiveMode(setting);
+      break;
+    case 3:
+      setting = importSetting();
+      singleConversionMode(setting);
+      break;
+    case 4:
+      if (process.argv[2] == "setting") {
+        if (process.argv[3] == "init") {
+          makeSettingFile();
+        }
+      }
+      break;
+    default:
+      console.log("Invalid parameter.");
+      break;
   }
 };
