+++ lib/scripts/single-conversion-mode.js
@@ -1,0 +1,8 @@
+const { translateText } = require("../util/translate-text");
+
+exports.singleConversionMode = async (setting) => {
+  const translatedText = await translateText(process.argv[2], setting);
+
+  console.log(translatedText);
+  process.exit(0);
+};
